requirements.txt
selenium
webdriver-manager
pypdf
mysql-connector-python
pandas
pytest 
flake8 
black